deliver amountwater to higher lvl
