+++ src/components/trackerPage/WaterDetailedInfo/WaterDetailedInfo.jsx
@@ -3,11 +3,11 @@ import { MonthInfo } from '../MonthInfo/MonthInfo';
 import { UserPanel } from '../UserPanel/UserPanel';
 import s from './WaterDetailedInfo.module.css';
 
-const WaterDetailedInfo = () => {
+const WaterDetailedInfo = ({amountData, setAmountData, isError, isLoading}) => {
   return (
     <div className={s.waterDetailedInfo}>
       <UserPanel />
-      <DailyInfo />
+      <DailyInfo amountData={amountData} setAmountData={setAmountData} isError={isError} isLoading={isLoading} />
       <MonthInfo />
     </div>
   );
